requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
GDAL==3.2.2
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.0
numpy==1.19.0
osgeo==0.0.0
pandas==1.0.5
pep8==1.7.1
Pillow==7.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.24.0
scipy==1.5.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.0
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
whitenoise==5.0.1
zope.interface==5.1.2
